requirements.txt
django
django-db
pytest
pytest-django
mock
django-rest-framework
requests
coreapi
drf-yasg